server/requirements.txt
Django==1.10.1
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.2
django-compressor==1.4
django-debug-toolbar==1.4
django-toolbelt==0.0.1
djangorestframework==3.4.7
drf-nested-routers==0.9.0
gunicorn==19.1.1
psycopg2==2.5.4
six==1.8.0
sqlparse==0.1.13
static3==0.5.1
wsgiref==0.1.2
django-cors-headers==1.1.0
requests==2.11.1